split components out of App.js
